+++ src/fast_test.py
@@ -1,0 +1,8 @@
+#! /usr/bin/env python3
+
+import unittest
+
+from test import TestGeneratorGECCO
+
+if __name__ == '__main__':
+    unittest.main()